chore: DOM 조작 함수 이동 DOM이 사용되는 컴포넌트 위치로 함수 이동
