+++ src/components/MovieList.ts
@@ -6,24 +6,8 @@ import { $, $$ } from '../utils';
 
 import { MovieListSkeleton } from './MovieListSkeleton';
 import { MovieItem } from './MovieItem';
-import { PAGE_TITLE } from '../constants/constants';
-
-export function renderPageTitle(keyword: string, isPopular: boolean) {
-  const $pageTitle = $('.page-title') as HTMLElement;
-  if (!isPopular && keyword) {
-    $pageTitle.innerText = PAGE_TITLE.showSearchResult(keyword);
-    return;
-  }
-
-  $pageTitle.innerText = PAGE_TITLE.POPULAR_NOW;
-}
-
-export function renderViewMoreButton(isContentEnd: boolean) {
-  if (isContentEnd) {
-    const $viewMoreButton = $('.view-more-button') as HTMLElement;
-    $viewMoreButton.style.display = 'none';
-  }
-}
+import { renderPageTitle } from './PageTitle';
+import { renderViewMoreButton } from './ViewMoreButton';
 
 export async function renderSkeletonList(state: publisher) {
   const { page, keyword, isPopular } = state;
@@ -44,17 +28,16 @@ export async function renderSkeletonList(state: publisher) {
 
   renderSearchMovieList(results);
   renderViewMoreButton(results.length < 20);
-  renderMoreMovieList(results);
 }
 
 export async function renderMoreSkeletonList(state: publisher) {
-  const { keyword, isPopular } = state;
+  const { page, keyword, isPopular } = state;
 
   const parentElem = $('.item-list') as HTMLElement;
   parentElem.insertAdjacentHTML('beforeend', MovieListSkeleton());
 
   if (isPopular) {
-    const { results } = await getPopularMovies(state.page + 1);
+    const { results } = await getPopularMovies(page + 1);
 
     renderMoreMovieList(results);
     deleteSkeletonList();
@@ -62,25 +45,19 @@ export async function renderMoreSkeletonList(state: publisher) {
     return;
   }
 
-  const { results } = await getSearchMovie(keyword, state.page + 1);
+  const { results } = await getSearchMovie(keyword, page + 1);
 
-  renderViewMoreButton(results.length < 20);
   renderMoreMovieList(results);
+  renderViewMoreButton(results.length < 20);
 
   deleteSkeletonList();
-}
-
-export function deleteSkeletonList() {
-  const skeletonList = $$('.skeleton-item');
-
-  skeletonList?.forEach((item) => item.remove());
 }
 
 export async function renderPopularMovieList(results: IMovie[]) {
   const parentElem = $('.item-list') as HTMLElement;
 
   parentElem.innerHTML = `
-    ${results.map((movie) => MovieItem(movie)).join('')}
+  ${results.map((movie) => MovieItem(movie)).join('')}
   
   `;
 }
@@ -92,7 +69,7 @@ export async function renderSearchMovieList(searchResults: IMovie[]) {
     searchResults.length === 0
       ? '검색 결과가 없습니다.'
       : `
-    ${searchResults.map((movie) => MovieItem(movie)).join('')}
+  ${searchResults.map((movie) => MovieItem(movie)).join('')}
   
   `;
 }
@@ -104,6 +81,12 @@ export async function renderMoreMovieList(moreResults: IMovie[]) {
     'beforeend',
     `${moreResults.map((movie) => MovieItem(movie)).join('')}`
   );
+}
+
+export function deleteSkeletonList() {
+  const skeletonList = $$('.skeleton-item');
+
+  skeletonList?.forEach((item) => item.remove());
 }
 
 export function MovieList() {
